Add files via upload
These files are datasets extrapolated from the OECD database and UNData. The dataset extrapolated from the OECD database looks at the percentage of families urbanised in different countries while the UNData dataset examines all countries gender inequality index. I want to look at the relationship between urbanisation and gender equality and how much it impacts it. The inspiration for my project was this article https://www.economist.com/international/2021/09/11/societies-that-treat-women-badly-are-poorer-and-less-stable